refactor(test): P5-2 extract CG-paths cluster from test_canvas
Phase 5 P5-2 follow-up. Apple-only CoreGraphicsCanvas tests for
Canvas2D paths / transforms / gradients / blend / save/restore
extracted to test_canvas_cg_paths.cpp (12 TEST_CASEs, 148 assertions,
~800 lines). Covers:

  * pulp #943 / #933 concat_transform (translates + scales)
  * pulp #1322 Canvas2D path API (fill, quad/cubic curves, stroke)
  * pulp #1322 linear gradient paint + fill_rect/path with active gradient
  * pulp #1368 save_count + restore_to_count tracking
  * pulp #1322 set_blend_mode every-enum-value round-trip
  * Canvas default save_count/restore_to_count (moved with MinimalCanvas)

Companion to test_canvas_cg_gradients.cpp (PR #2247). The cluster
shares MinimalCanvas + Apple framework links; the parent TU keeps the
non-Apple paths only.

test_canvas.cpp: 2,026 → 1,254 (-772 lines, under 2k P5-2 target)
