+++ test/test_canvas.cpp
@@ -935,779 +935,7 @@ TEST_CASE("CoreGraphicsCanvas::clear_rect zeroes destination pixels",
         REQUIRE(pixels[i] == 0);
     }
 }
-
-// pulp #943 (#933 P1) — CoreGraphicsCanvas::concat_transform must compose the
-// supplied affine onto the current CTM rather than no-op (the default in the
-// base Canvas virtual). Without the override, View::paint_all() routes
-// JS-supplied setTransform(...) through Canvas::concat_transform and the
-// transform silently disappears on Apple CPU paint paths.
-//
-// Strategy: render a marker rectangle with two equivalent code paths and
-// require the bitmaps come out byte-identical:
-//   (a) no concat_transform, draw at (50 + dx, dy)
-//   (b) concat_transform(1, 0, 0, 1, 50, 0), draw at (dx, dy)
-// If concat_transform is a no-op (the bug), path (b) draws at (dx, dy) and
-// the bitmaps differ. With the override calling CGContextConcatCTM, both
-// paths land at the same destination pixels.
-TEST_CASE("CoreGraphicsCanvas::concat_transform translates draw position",
-          "[canvas][cg][issue-943-933]") {
-    constexpr int W = 64;
-    constexpr int H = 32;
-    auto build_ctx = [&](std::vector<uint8_t>& pixels) -> CGContextRef {
-        pixels.assign(static_cast<size_t>(W) * H * 4u, 0u);
-        auto cs = CGColorSpaceCreateDeviceRGB();
-        REQUIRE(cs != nullptr);
-        const uint32_t bitmap_info =
-            static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-            static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-        CGContextRef ctx = CGBitmapContextCreate(
-            pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-        CGColorSpaceRelease(cs);
-        REQUIRE(ctx != nullptr);
-        return ctx;
-    };
-
-    // (a) Reference render — draw a 4x4 red rect at canvas (60, 10) directly.
-    std::vector<uint8_t> reference(static_cast<size_t>(W) * H * 4u, 0u);
-    {
-        CGContextRef ctx = build_ctx(reference);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            canvas.set_fill_color(Color::rgba(1.0f, 0.0f, 0.0f, 1.0f));
-            canvas.fill_rect(60.0f, 10.0f, 4.0f, 4.0f);
-        }
-        CGContextRelease(ctx);
-    }
-
-    // (b) Through concat_transform — translate +50 in x, then draw at (10, 10).
-    // With the override in place this must produce the same final pixels.
-    std::vector<uint8_t> via_concat(static_cast<size_t>(W) * H * 4u, 0u);
-    {
-        CGContextRef ctx = build_ctx(via_concat);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            // Pure translation: a=1, b=0, c=0, d=1, e=50, f=0.
-            canvas.concat_transform(1.0f, 0.0f, 0.0f, 1.0f, 50.0f, 0.0f);
-            canvas.set_fill_color(Color::rgba(1.0f, 0.0f, 0.0f, 1.0f));
-            canvas.fill_rect(10.0f, 10.0f, 4.0f, 4.0f);
-        }
-        CGContextRelease(ctx);
-    }
-
-    REQUIRE(reference.size() == via_concat.size());
-    REQUIRE(reference == via_concat);
-
-    // Sanity: there must be a non-trivial number of red pixels — guards
-    // against a regression that no-ops both paths and produces empty bitmaps.
-    int red_pixels = 0;
-    for (size_t i = 0; i + 3 < reference.size(); i += 4) {
-        if (reference[i] == 255 && reference[i + 1] == 0 &&
-            reference[i + 2] == 0 && reference[i + 3] == 255) {
-            ++red_pixels;
-        }
-    }
-    REQUIRE(red_pixels >= 16);  // at least the 4x4 footprint
-}
-
-// pulp #943 (#933 P1) — verify a non-translation affine (scale + translate)
-// composes correctly, not just pure translations. This catches a regression
-// where someone implements concat_transform as CGContextTranslateCTM(e, f)
-// and ignores a/b/c/d.
-TEST_CASE("CoreGraphicsCanvas::concat_transform scales + translates",
-          "[canvas][cg][issue-943-933]") {
-    constexpr int W = 64;
-    constexpr int H = 32;
-    auto build_ctx = [&](std::vector<uint8_t>& pixels) -> CGContextRef {
-        pixels.assign(static_cast<size_t>(W) * H * 4u, 0u);
-        auto cs = CGColorSpaceCreateDeviceRGB();
-        const uint32_t bitmap_info =
-            static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-            static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-        CGContextRef ctx = CGBitmapContextCreate(
-            pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-        CGColorSpaceRelease(cs);
-        REQUIRE(ctx != nullptr);
-        return ctx;
-    };
-
-    // Reference: scale 2x in x, then draw a 4x4 rect at (10, 10) — should
-    // cover canvas-space (20, 10)..(28, 14).
-    std::vector<uint8_t> reference(static_cast<size_t>(W) * H * 4u, 0u);
-    {
-        CGContextRef ctx = build_ctx(reference);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            canvas.scale(2.0f, 1.0f);
-            canvas.set_fill_color(Color::rgba(0.0f, 1.0f, 0.0f, 1.0f));
-            canvas.fill_rect(10.0f, 10.0f, 4.0f, 4.0f);
-        }
-        CGContextRelease(ctx);
-    }
-
-    // Via concat_transform with sx=2, sy=1, tx=ty=0.
-    std::vector<uint8_t> via_concat(static_cast<size_t>(W) * H * 4u, 0u);
-    {
-        CGContextRef ctx = build_ctx(via_concat);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            canvas.concat_transform(2.0f, 0.0f, 0.0f, 1.0f, 0.0f, 0.0f);
-            canvas.set_fill_color(Color::rgba(0.0f, 1.0f, 0.0f, 1.0f));
-            canvas.fill_rect(10.0f, 10.0f, 4.0f, 4.0f);
-        }
-        CGContextRelease(ctx);
-    }
-
-    REQUIRE(reference == via_concat);
-}
-
-// pulp #1322 — CoreGraphicsCanvas must implement Canvas2D-style path
-// building. The base Canvas defaults are no-ops, so a JS bundle that drives
-// draw via beginPath/moveTo/lineTo/closePath/fillPath silently produced
-// nothing on the CPU paint path used by Pulp's standalone host
-// (run_with_editor(use_gpu=false)), even though the bridge dutifully
-// dispatched ~1800 commands per frame. Spectr's FilterBank canvas is the
-// canonical repro — see issue thread.
-//
-// The test draws a 4-vertex diamond polygon via beginPath/moveTo/lineTo*3/
-// closePath/fill_current_path and asserts that filled red pixels actually
-// appear in the destination bitmap. Without the override the buffer stays
-// fully zero and the count of red pixels is 0.
-TEST_CASE("CoreGraphicsCanvas Canvas2D path API fills (issue 1322)",
-          "[canvas][cg][issue-1322]") {
-    constexpr int W = 32;
-    constexpr int H = 32;
-    std::vector<uint8_t> pixels(static_cast<size_t>(W) * H * 4u, 0u);
-    auto cs = CGColorSpaceCreateDeviceRGB();
-    REQUIRE(cs != nullptr);
-    const uint32_t bitmap_info =
-        static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-        static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-    CGContextRef ctx = CGBitmapContextCreate(
-        pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-    CGColorSpaceRelease(cs);
-    REQUIRE(ctx != nullptr);
-
-    {
-        CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                  static_cast<float>(H));
-        canvas.set_fill_color(Color::rgba(1.0f, 0.0f, 0.0f, 1.0f));
-        // Diamond covering the centre of the bitmap (16,8)→(24,16)→(16,24)→(8,16).
-        canvas.begin_path();
-        canvas.move_to(16.0f, 8.0f);
-        canvas.line_to(24.0f, 16.0f);
-        canvas.line_to(16.0f, 24.0f);
-        canvas.line_to(8.0f, 16.0f);
-        canvas.close_path();
-        canvas.fill_current_path();
-    }
-    CGContextRelease(ctx);
-
-    int red_pixels = 0;
-    for (size_t i = 0; i + 3 < pixels.size(); i += 4) {
-        if (pixels[i] >= 200 && pixels[i + 1] <= 60 &&
-            pixels[i + 2] <= 60 && pixels[i + 3] >= 200) {
-            ++red_pixels;
-        }
-    }
-    INFO("red_pixels=" << red_pixels);
-    REQUIRE(red_pixels >= 16);  // diamond covers ~64 pixels at this size
-}
-
-// pulp #1322 — beziers (quadTo, cubicTo) must also accumulate into the
-// Canvas2D path. Same shape coverage check as the diamond test.
-TEST_CASE("CoreGraphicsCanvas Canvas2D path quad/cubic curves fill (issue 1322)",
-          "[canvas][cg][issue-1322]") {
-    constexpr int W = 32;
-    constexpr int H = 32;
-    std::vector<uint8_t> pixels(static_cast<size_t>(W) * H * 4u, 0u);
-    auto cs = CGColorSpaceCreateDeviceRGB();
-    const uint32_t bitmap_info =
-        static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-        static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-    CGContextRef ctx = CGBitmapContextCreate(
-        pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-    CGColorSpaceRelease(cs);
-    REQUIRE(ctx != nullptr);
-
-    {
-        CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                  static_cast<float>(H));
-        canvas.set_fill_color(Color::rgba(0.0f, 0.0f, 1.0f, 1.0f));
-        canvas.begin_path();
-        canvas.move_to(8.0f, 16.0f);
-        canvas.quad_to(16.0f, 4.0f, 24.0f, 16.0f);
-        canvas.cubic_to(20.0f, 24.0f, 12.0f, 24.0f, 8.0f, 16.0f);
-        canvas.close_path();
-        canvas.fill_current_path();
-    }
-    CGContextRelease(ctx);
-
-    int blue_pixels = 0;
-    for (size_t i = 0; i + 3 < pixels.size(); i += 4) {
-        if (pixels[i] <= 60 && pixels[i + 1] <= 60 &&
-            pixels[i + 2] >= 200 && pixels[i + 3] >= 200) {
-            ++blue_pixels;
-        }
-    }
-    INFO("blue_pixels=" << blue_pixels);
-    REQUIRE(blue_pixels >= 16);
-}
-
-// pulp #1322 — Canvas2D path stroke must hit the destination too. Spectr
-// also draws spectrum traces via beginPath/moveTo/lineTo*N/strokePath, and
-// stroke_current_path was a no-op on CG before this fix.
-TEST_CASE("CoreGraphicsCanvas Canvas2D path stroke draws (issue 1322)",
-          "[canvas][cg][issue-1322]") {
-    constexpr int W = 32;
-    constexpr int H = 32;
-    std::vector<uint8_t> pixels(static_cast<size_t>(W) * H * 4u, 0u);
-    auto cs = CGColorSpaceCreateDeviceRGB();
-    const uint32_t bitmap_info =
-        static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-        static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-    CGContextRef ctx = CGBitmapContextCreate(
-        pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-    CGColorSpaceRelease(cs);
-    REQUIRE(ctx != nullptr);
-
-    {
-        CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                  static_cast<float>(H));
-        canvas.set_stroke_color(Color::rgba(0.0f, 1.0f, 0.0f, 1.0f));
-        canvas.set_line_width(2.0f);
-        canvas.begin_path();
-        canvas.move_to(2.0f, 16.0f);
-        canvas.line_to(30.0f, 16.0f);
-        canvas.stroke_current_path();
-    }
-    CGContextRelease(ctx);
-
-    int green_pixels = 0;
-    for (size_t i = 0; i + 3 < pixels.size(); i += 4) {
-        if (pixels[i] <= 60 && pixels[i + 1] >= 200 &&
-            pixels[i + 2] <= 60 && pixels[i + 3] >= 200) {
-            ++green_pixels;
-        }
-    }
-    INFO("green_pixels=" << green_pixels);
-    REQUIRE(green_pixels >= 24);  // a 28-pixel-wide line at width=2
-}
-
-// pulp #1322 — set_fill_gradient_linear must paint a real gradient on CG;
-// the base Canvas fallback collapses to "set fill color = colors[0]" which
-// produces a single colour fill (Spectr's spectrum bg is gradient-driven).
-// Verify that two different colors actually appear in the output bitmap.
-TEST_CASE("CoreGraphicsCanvas linear gradient paints multiple colors (issue 1322)",
-          "[canvas][cg][issue-1322]") {
-    constexpr int W = 64;
-    constexpr int H = 16;
-    std::vector<uint8_t> pixels(static_cast<size_t>(W) * H * 4u, 0u);
-    auto cs = CGColorSpaceCreateDeviceRGB();
-    const uint32_t bitmap_info =
-        static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-        static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-    CGContextRef ctx = CGBitmapContextCreate(
-        pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-    CGColorSpaceRelease(cs);
-    REQUIRE(ctx != nullptr);
-
-    {
-        CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                  static_cast<float>(H));
-        Color colors[2] = {
-            Color::rgba(1.0f, 0.0f, 0.0f, 1.0f),
-            Color::rgba(0.0f, 0.0f, 1.0f, 1.0f),
-        };
-        float positions[2] = {0.0f, 1.0f};
-        canvas.set_fill_gradient_linear(0, 0, static_cast<float>(W), 0,
-                                         colors, positions, 2);
-        canvas.fill_rect(0, 0, static_cast<float>(W), static_cast<float>(H));
-    }
-    CGContextRelease(ctx);
-
-    bool saw_red_dominant = false;
-    bool saw_blue_dominant = false;
-    for (size_t i = 0; i + 3 < pixels.size(); i += 4) {
-        if (pixels[i] >= 200 && pixels[i + 2] <= 60) saw_red_dominant = true;
-        if (pixels[i] <= 60 && pixels[i + 2] >= 200) saw_blue_dominant = true;
-    }
-    REQUIRE(saw_red_dominant);
-    REQUIRE(saw_blue_dominant);
-}
-
-// pulp #1359 — fill_rect already routes the active gradient through
-// fill_with_active_paint(), but fill_path / fill_circle / fill_rounded_rect
-// silently dropped the gradient and fell back to apply_fill_color(). This
-// is the direct CG parallel of pulp #1350/#1353 on the Skia side. Spectr's
-// CPU-mode FilterBank backplate is the canonical repro — it paints solid
-// white instead of the dark-gradient backplate without this fix.
-//
-// Verify a 64x8 rect filled with a red→green linear gradient produces a
-// red-dominant left endpoint and a green-dominant right endpoint.
-TEST_CASE("CoreGraphicsCanvas::fill_rect honors active linear gradient",
-          "[canvas][cg][gradient][issue-1359]") {
-    constexpr int W = 64;
-    constexpr int H = 8;
-    std::vector<uint8_t> pixels(static_cast<size_t>(W) * H * 4u, 0u);
-    auto cs = CGColorSpaceCreateDeviceRGB();
-    REQUIRE(cs != nullptr);
-    const uint32_t bitmap_info =
-        static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-        static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-    CGContextRef ctx = CGBitmapContextCreate(
-        pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-    CGColorSpaceRelease(cs);
-    REQUIRE(ctx != nullptr);
-
-    {
-        CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                  static_cast<float>(H));
-        Color colors[2] = {
-            Color::rgba(1.0f, 0.0f, 0.0f, 1.0f),
-            Color::rgba(0.0f, 1.0f, 0.0f, 1.0f),
-        };
-        float positions[2] = {0.0f, 1.0f};
-        canvas.set_fill_gradient_linear(0, 0, static_cast<float>(W), 0,
-                                         colors, positions, 2);
-        canvas.fill_rect(0, 0, static_cast<float>(W), static_cast<float>(H));
-    }
-    CGContextRelease(ctx);
-
-    // Sample two pixels — one near the left edge, one near the right edge,
-    // both on a middle row to dodge any antialiasing along the top/bottom.
-    auto sample = [&](int x, int y) {
-        const size_t idx = (static_cast<size_t>(y) * W + x) * 4u;
-        return std::tuple<int, int, int, int>{pixels[idx], pixels[idx + 1],
-                                              pixels[idx + 2], pixels[idx + 3]};
-    };
-    auto [lr, lg, lb, la] = sample(2, H / 2);
-    auto [rr, rg, rb, rb_a] = sample(W - 3, H / 2);
-    INFO("left rgba=" << lr << "," << lg << "," << lb << "," << la);
-    INFO("right rgba=" << rr << "," << rg << "," << rb << "," << rb_a);
-    // Directional check: left endpoint must be more red than green; right
-    // endpoint must be more green than red. Tolerant of CG color-space drift.
-    REQUIRE(lr > lg);
-    REQUIRE(rg > rr);
-}
-
-// pulp #1359 — same test for fill_path. Build a triangle path and verify
-// at least two pixels inside the rendered triangle differ in color, proving
-// the gradient was actually painted (not a single solid colour).
-TEST_CASE("CoreGraphicsCanvas::fill_path honors active linear gradient",
-          "[canvas][cg][gradient][issue-1359]") {
-    constexpr int W = 64;
-    constexpr int H = 32;
-    std::vector<uint8_t> pixels(static_cast<size_t>(W) * H * 4u, 0u);
-    auto cs = CGColorSpaceCreateDeviceRGB();
-    REQUIRE(cs != nullptr);
-    const uint32_t bitmap_info =
-        static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-        static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-    CGContextRef ctx = CGBitmapContextCreate(
-        pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-    CGColorSpaceRelease(cs);
-    REQUIRE(ctx != nullptr);
-
-    {
-        CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                  static_cast<float>(H));
-        Color colors[2] = {
-            Color::rgba(1.0f, 0.0f, 0.0f, 1.0f),
-            Color::rgba(0.0f, 1.0f, 0.0f, 1.0f),
-        };
-        float positions[2] = {0.0f, 1.0f};
-        canvas.set_fill_gradient_linear(0, 0, static_cast<float>(W), 0,
-                                         colors, positions, 2);
-
-        // Triangle covering most of the bitmap horizontally so the gradient
-        // is sampled across its full extent.
-        Canvas::Point2D tri[3] = {
-            {4.0f, 4.0f},
-            {static_cast<float>(W) - 4.0f, static_cast<float>(H) / 2.0f},
-            {4.0f, static_cast<float>(H) - 4.0f},
-        };
-        canvas.fill_path(tri, 3);
-    }
-    CGContextRelease(ctx);
-
-    // Walk the bitmap and count pixels that look red-dominant vs
-    // green-dominant. If fill_path dropped the gradient and fell back to
-    // apply_fill_color(), every painted pixel would be a single colour and
-    // exactly one of these counts would be non-zero.
-    int red_dominant = 0;
-    int green_dominant = 0;
-    for (size_t i = 0; i + 3 < pixels.size(); i += 4) {
-        const int r = pixels[i];
-        const int g = pixels[i + 1];
-        const int a = pixels[i + 3];
-        if (a < 10) continue;  // background
-        if (r >= 150 && g <= 60) ++red_dominant;
-        if (g >= 150 && r <= 60) ++green_dominant;
-    }
-    INFO("red_dominant=" << red_dominant
-         << " green_dominant=" << green_dominant);
-    REQUIRE(red_dominant > 0);
-    REQUIRE(green_dominant > 0);
-}
-
-// pulp #1368 — CoreGraphicsCanvas tracks save_count() and supports
-// restore_to_count() for the CanvasWidget::paint defensive bracket.
-TEST_CASE("CoreGraphicsCanvas tracks save_count and restore_to_count",
-          "[canvas][cg][issue-1368]") {
-    constexpr int W = 16;
-    constexpr int H = 16;
-    auto build_ctx = [&](std::vector<uint8_t>& pixels) {
-        auto colorSpace = CGColorSpaceCreateDeviceRGB();
-        CGContextRef ctx = CGBitmapContextCreate(
-            pixels.data(), static_cast<size_t>(W), static_cast<size_t>(H),
-            8, static_cast<size_t>(W) * 4u, colorSpace,
-            kCGImageAlphaPremultipliedLast | kCGBitmapByteOrder32Big);
-        CGColorSpaceRelease(colorSpace);
-        return ctx;
-    };
-    std::vector<uint8_t> pixels(static_cast<size_t>(W) * H * 4u, 0u);
-    CGContextRef ctx = build_ctx(pixels);
-    {
-        CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                  static_cast<float>(H));
-        const int baseline = canvas.save_count();
-
-        canvas.save();
-        canvas.save();
-        canvas.save();
-        REQUIRE(canvas.save_count() == baseline + 3);
-
-        // Pop two levels — depth drops by exactly two.
-        canvas.restore_to_count(baseline + 1);
-        REQUIRE(canvas.save_count() == baseline + 1);
-
-        // restore_to_count to original baseline drops the last leftover.
-        canvas.restore_to_count(baseline);
-        REQUIRE(canvas.save_count() == baseline);
-
-        // restore_to_count below baseline is a clamped no-op (no underflow).
-        canvas.restore_to_count(baseline - 5);
-        REQUIRE(canvas.save_count() == baseline);
-    }
-    CGContextRelease(ctx);
-}
-
-// pulp #1371 — CoreGraphicsCanvas::set_blend_mode was a silent no-op (the
-// base Canvas virtual default `(void)mode;`). Skia honored every CSS
-// globalCompositeOperation; CG dropped them all and forced SrcOver. The
-// canonical repro is Spectr's filterbank: `ctx.globalCompositeOperation =
-// 'lighter'` paints a vivid blue→green→red rainbow gradient additively over
-// the dark canvas; without the override the gradient barely tinted the
-// backplate.
-//
-// Strategy: set up a CG bitmap context, paint an opaque red base layer, then
-// fill a second rect of equal extent with a different blend mode and assert
-// that the resulting destination pixel matches the chosen op's spec, NOT the
-// SrcOver default. Three coverage points:
-//
-//   * `BlendMode::multiply` — red(255,0,0) * blue(0,0,255) ≈ black; under
-//     SrcOver the dest would be plain blue.
-//   * `BlendMode::lighter` — kCGBlendModePlusLighter; the result must be
-//     strictly brighter than either input on every channel that contributed.
-//   * `BlendMode::copy` — kCGBlendModeCopy; replaces destination outright,
-//     proving non-default Porter-Duff modes also reach CG.
-//   * `BlendMode::xor_mode` — kCGBlendModeXOR; covers an additional CSS
-//     keyword path (issue-896 surface).
-TEST_CASE("CoreGraphicsCanvas::set_blend_mode honors all BlendMode values",
-          "[canvas][cg][blend][issue-1371]") {
-    constexpr int W = 8;
-    constexpr int H = 8;
-    auto build_ctx = [&](std::vector<uint8_t>& pixels) -> CGContextRef {
-        pixels.assign(static_cast<size_t>(W) * H * 4u, 0u);
-        auto cs = CGColorSpaceCreateDeviceRGB();
-        REQUIRE(cs != nullptr);
-        const uint32_t bitmap_info =
-            static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-            static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-        CGContextRef ctx = CGBitmapContextCreate(
-            pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-        CGColorSpaceRelease(cs);
-        REQUIRE(ctx != nullptr);
-        return ctx;
-    };
-
-    // Pixel at (W/2, H/2) is the center of the painted area — every test
-    // fills (0,0,W,H) twice so the center is always inside both rects.
-    auto sample_center = [&](const std::vector<uint8_t>& pixels) {
-        const size_t row = (H / 2);
-        const size_t col = (W / 2);
-        const size_t idx = (row * W + col) * 4u;
-        struct RGBA { uint8_t r, g, b, a; };
-        return RGBA{pixels[idx + 0], pixels[idx + 1],
-                    pixels[idx + 2], pixels[idx + 3]};
-    };
-
-    SECTION("multiply — red × blue ≈ black") {
-        std::vector<uint8_t> pixels;
-        CGContextRef ctx = build_ctx(pixels);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            canvas.set_fill_color(Color::rgba(1.0f, 0.0f, 0.0f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-            canvas.set_blend_mode(Canvas::BlendMode::multiply);
-            canvas.set_fill_color(Color::rgba(0.0f, 0.0f, 1.0f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-        }
-        CGContextRelease(ctx);
-
-        auto px = sample_center(pixels);
-        // multiply(red, blue) per-channel: r*0=0, g*0=0, b*0=0 — pure black.
-        // Under SrcOver this would be (0,0,255) — plain blue. The bug fix
-        // gates on r AND b both being zero.
-        INFO("center pixel rgba=(" << int(px.r) << "," << int(px.g)
-             << "," << int(px.b) << "," << int(px.a) << ")");
-        REQUIRE(px.r < 8);
-        REQUIRE(px.g < 8);
-        REQUIRE(px.b < 8);
-        REQUIRE(px.a == 255);
-    }
-
-    SECTION("lighter (kCGBlendModePlusLighter) — additive sum") {
-        std::vector<uint8_t> pixels;
-        CGContextRef ctx = build_ctx(pixels);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            // Half-strength red on black background.
-            canvas.set_fill_color(Color::rgba(0.5f, 0.0f, 0.0f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-            canvas.set_blend_mode(Canvas::BlendMode::lighter);
-            // Add half-strength green — sum should be (0.5, 0.5, 0).
-            canvas.set_fill_color(Color::rgba(0.0f, 0.5f, 0.0f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-        }
-        CGContextRelease(ctx);
-
-        auto px = sample_center(pixels);
-        INFO("center pixel rgba=(" << int(px.r) << "," << int(px.g)
-             << "," << int(px.b) << "," << int(px.a) << ")");
-        // lighter is additive — both red AND green channels must be
-        // present. Under SrcOver the second fill would replace red with
-        // pure green (r=0, g=128) — additive must keep red ≈ 128 too.
-        REQUIRE(px.r >= 96);   // ~0.5 * 255 = 128, allow rounding slack
-        REQUIRE(px.g >= 96);
-        REQUIRE(px.b < 16);
-    }
-
-    SECTION("copy (kCGBlendModeCopy) — replaces destination") {
-        std::vector<uint8_t> pixels;
-        CGContextRef ctx = build_ctx(pixels);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            canvas.set_fill_color(Color::rgba(1.0f, 0.0f, 0.0f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-            canvas.set_blend_mode(Canvas::BlendMode::copy);
-            // Half-transparent green — under copy the destination becomes
-            // exactly this premultiplied source, NOT the SrcOver blend.
-            canvas.set_fill_color(Color::rgba(0.0f, 1.0f, 0.0f, 0.5f));
-            canvas.fill_rect(0, 0, W, H);
-        }
-        CGContextRelease(ctx);
-
-        auto px = sample_center(pixels);
-        INFO("center pixel rgba=(" << int(px.r) << "," << int(px.g)
-             << "," << int(px.b) << "," << int(px.a) << ")");
-        // copy mode: destination = source. Red base must be gone (SrcOver
-        // would leave red ≈ 128 from blending with the half-alpha green).
-        REQUIRE(px.r < 16);
-        REQUIRE(px.a < 200);  // alpha is the half-alpha source, not 255
-    }
-
-    SECTION("xor_mode — issue-896 CSS keyword path reaches CG") {
-        std::vector<uint8_t> pixels;
-        CGContextRef ctx = build_ctx(pixels);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            canvas.set_fill_color(Color::rgba(1.0f, 0.0f, 0.0f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-            canvas.set_blend_mode(Canvas::BlendMode::xor_mode);
-            // Same opaque red on top — XOR of two opaque solids is fully
-            // transparent in spec, regardless of color.
-            canvas.set_fill_color(Color::rgba(1.0f, 0.0f, 0.0f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-        }
-        CGContextRelease(ctx);
-
-        auto px = sample_center(pixels);
-        INFO("center pixel rgba=(" << int(px.r) << "," << int(px.g)
-             << "," << int(px.b) << "," << int(px.a) << ")");
-        // XOR of two fully-opaque (a=1) overlapping solids → alpha 0.
-        // Under SrcOver this would be opaque red (a=255).
-        REQUIRE(px.a < 32);
-    }
-
-    SECTION("normal (default) sanity — SrcOver still works after fix") {
-        // Guards against a regression where the new override broke the
-        // default path. With normal, an opaque blue rect drawn on top of
-        // an opaque red rect must produce blue.
-        std::vector<uint8_t> pixels;
-        CGContextRef ctx = build_ctx(pixels);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            canvas.set_fill_color(Color::rgba(1.0f, 0.0f, 0.0f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-            canvas.set_blend_mode(Canvas::BlendMode::normal);
-            canvas.set_fill_color(Color::rgba(0.0f, 0.0f, 1.0f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-        }
-        CGContextRelease(ctx);
-
-        auto px = sample_center(pixels);
-        INFO("center pixel rgba=(" << int(px.r) << "," << int(px.g)
-             << "," << int(px.b) << "," << int(px.a) << ")");
-        REQUIRE(px.r < 16);
-        REQUIRE(px.g < 16);
-        REQUIRE(px.b > 240);
-        REQUIRE(px.a == 255);
-    }
-}
-
-// pulp #1371 — exhaustively exercise every BlendMode enum case in the new
-// switch so the diff-cover gate sees each branch run. The earlier test
-// proves end-to-end pixel correctness on a handful of representative ops;
-// this one is structural — every enum value must round-trip through
-// `CoreGraphicsCanvas::set_blend_mode → to_cg_blend(...)` and reach CG.
-//
-// We don't assert per-channel pixel formulas for every mode (CG's edge
-// behavior for hue / saturation / color / luminosity at the bitmap-context
-// level depends on Apple's internal LUTs and would be brittle). The
-// invariant we assert instead: applying the blend mode and painting must
-// not crash the CG context, and the result must be reproducible (no
-// undefined behaviour). For most modes the result is non-empty; for
-// `source_out` / `destination_out` (where the result IS empty when
-// source and destination cover the same area) we whitelist that as the
-// CSS-spec behaviour. Coverage hits every case branch in the switch.
-TEST_CASE("CoreGraphicsCanvas::set_blend_mode every enum value round-trips through to_cg_blend",
-          "[canvas][cg][blend][issue-1371]") {
-    constexpr int W = 4;
-    constexpr int H = 4;
-    using BM = Canvas::BlendMode;
-    // Every enum value listed in canvas.hpp BlendMode in declaration order.
-    const std::vector<BM> all_modes{
-        BM::normal,        BM::multiply,    BM::screen,        BM::overlay,
-        BM::darken,        BM::lighten,     BM::color_dodge,   BM::color_burn,
-        BM::hard_light,    BM::soft_light,  BM::difference,    BM::exclusion,
-        BM::hue,           BM::saturation,  BM::color,         BM::luminosity,
-        BM::source_over,   BM::destination_over,
-        BM::source_in,     BM::destination_in,
-        BM::source_out,    BM::destination_out,
-        BM::source_atop,   BM::destination_atop,
-        BM::xor_mode,      BM::copy,        BM::lighter,
-    };
-
-    // Spec-empty modes: when source and destination cover the same area,
-    // the result is "destination minus source" or "source where destination
-    // isn't there" or "non-overlapping union" — all empty when the rects
-    // are fully coincident.
-    auto spec_allows_empty = [](BM m) {
-        return m == BM::source_out || m == BM::destination_out
-            || m == BM::xor_mode;
-    };
-
-    for (auto mode : all_modes) {
-        std::vector<uint8_t> pixels(static_cast<size_t>(W) * H * 4u, 0u);
-        auto cs = CGColorSpaceCreateDeviceRGB();
-        REQUIRE(cs != nullptr);
-        const uint32_t bitmap_info =
-            static_cast<uint32_t>(kCGImageAlphaPremultipliedLast) |
-            static_cast<uint32_t>(kCGBitmapByteOrder32Big);
-        CGContextRef ctx = CGBitmapContextCreate(
-            pixels.data(), W, H, 8, W * 4u, cs, bitmap_info);
-        CGColorSpaceRelease(cs);
-        REQUIRE(ctx != nullptr);
-        {
-            CoreGraphicsCanvas canvas(ctx, static_cast<float>(W),
-                                      static_cast<float>(H));
-            // Lay down a base layer the dest-side modes can interact with.
-            canvas.set_fill_color(Color::rgba(0.6f, 0.4f, 0.2f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-            canvas.set_blend_mode(mode);
-            canvas.set_fill_color(Color::rgba(0.2f, 0.5f, 0.7f, 1.0f));
-            canvas.fill_rect(0, 0, W, H);
-        }
-        CGContextRelease(ctx);
-        // Sample the centre pixel — this is post-blend output.
-        const size_t idx = (static_cast<size_t>(H / 2) * W + (W / 2)) * 4u;
-        const int r = pixels[idx + 0];
-        const int g = pixels[idx + 1];
-        const int b = pixels[idx + 2];
-        const int a = pixels[idx + 3];
-        INFO("mode=" << static_cast<int>(mode)
-             << " rgba=(" << r << "," << g << "," << b << "," << a << ")");
-        if (spec_allows_empty(mode)) {
-            // Empty result is correct (and what CG produces). Just confirm
-            // the values are deterministically inside [0,255]. The act of
-            // running the lambda body is what diff-cover counts, so this
-            // path still hits the case branch.
-            REQUIRE(r >= 0); REQUIRE(r <= 255);
-            REQUIRE(a >= 0); REQUIRE(a <= 255);
-        } else {
-            REQUIRE((r + g + b + a) > 0);
-        }
-    }
-}
-
 #endif  // __APPLE__
-
-// pulp #1368 — Canvas's default save_count() / restore_to_count() impls
-// are no-op fallbacks for backends that don't implement an introspectable
-// save stack. CanvasWidget's defensive bracket relies on the contract that
-// these are safe to call on any Canvas. Exercise the defaults via a
-// minimal Canvas subclass that doesn't override either method.
-namespace {
-
-class MinimalCanvas final : public pulp::canvas::Canvas {
-public:
-    void save() override {}
-    void restore() override {}
-    void translate(float, float) override {}
-    void scale(float, float) override {}
-    void rotate(float) override {}
-    void clip_rect(float, float, float, float) override {}
-    void set_fill_color(pulp::canvas::Color) override {}
-    void set_stroke_color(pulp::canvas::Color) override {}
-    void set_line_width(float) override {}
-    void set_line_cap(pulp::canvas::LineCap) override {}
-    void set_line_join(pulp::canvas::LineJoin) override {}
-    void fill_rect(float, float, float, float) override {}
-    void stroke_rect(float, float, float, float) override {}
-    void fill_rounded_rect(float, float, float, float, float) override {}
-    void stroke_rounded_rect(float, float, float, float, float) override {}
-    void fill_circle(float, float, float) override {}
-    void stroke_circle(float, float, float) override {}
-    void stroke_arc(float, float, float, float, float) override {}
-    void stroke_line(float, float, float, float) override {}
-    void set_font(const std::string&, float) override {}
-    void set_text_align(pulp::canvas::TextAlign) override {}
-    void fill_text(const std::string&, float, float) override {}
-    float measure_text(const std::string&) override { return 0.0f; }
-};
-
-} // namespace
-
-TEST_CASE("Canvas default save_count is 0 and restore_to_count is a no-op",
-          "[canvas][issue-1368]") {
-    MinimalCanvas mc;
-    REQUIRE(mc.save_count() == 0);
-    mc.restore_to_count(0);
-    mc.restore_to_count(-3);
-    mc.restore_to_count(99);
-    REQUIRE(mc.save_count() == 0);
-}
-
 
 // ── pulp #1737 — CSS font-variant → SkShaper Feature plumbing ─────────────
 // Regression coverage for the SkShaper 8-arg shape() overload that the
